move adapters to presentation layer and delete unUsed classes.
