--- app/src/main/java/com/nima/tmdb/adapters/LoadingViewHolder.kt
@@ -1,6 +1,0 @@
-package com.nima.tmdb.adapters
-
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-class LoadingViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
